完成项目前后端代码检查

+++ lease/model/src/main/java/com/zfd/lease/model/entity/CityInfo.java
@@ -18,6 +18,6 @@ public class CityInfo extends BaseEntity {
 
     @Schema(description = "所属省份id")
     @TableField(value = "province_id")
-    private Integer provinceId;
+    private Long provinceId;
 
 }
